requirements.txt
duckdb==1.0.0
httpx==0.27.2
python-dotenv==1.0.1
pyarrow==17.0.0
pytz==2024.1